+++ skeliner/io.py
@@ -115,8 +115,8 @@ def load_swc(
         raise ValueError(f"No usable nodes found in {path}")
 
     # --- core arrays ----------------------------------------------------
-    nodes_arr  = np.asarray(xyz, dtype=np.float32) * scale
-    radii_arr  = np.asarray(radii, dtype=np.float32) * scale
+    nodes_arr  = np.asarray(xyz, dtype=np.float64) * scale
+    radii_arr  = np.asarray(radii, dtype=np.float64) * scale
     radii_dict = {"median": radii_arr, "mean": radii_arr, "trim": radii_arr} 
     ntype_arr = np.asarray(ntype, dtype=np.int8)
     # --- edges (parent IDs → 0-based indices) ---------------------------
@@ -228,12 +228,12 @@ def load_npz(path: str | Path) -> Skeleton:
     path = Path(path)
 
     with np.load(path, allow_pickle=True) as z:
-        nodes  = z["nodes"].astype(np.float32)
+        nodes  = z["nodes"].astype(np.float64)
         edges  = z["edges"].astype(np.int64)
 
         # radii dict  (keys start with 'r_')
         radii = {
-            k[2:]: z[k].astype(np.float32) for k in z.files if k.startswith("r_")
+            k[2:]: z[k].astype(np.float64) for k in z.files if k.startswith("r_")
         }
 
         # node types (optional in older archives)
